This is the same bug fixed in Rev. 40, related to problems on 64 versus 32 bit arch.
Corrected bugs related to reading and writing binary files on 32 and 64 bit machines.
Binary hydraulics file was read using 'int' whereas it was writting using 'long' (INT4);
this was OK on 32 bit machines but on 64 bit machines caused an error. Similarly,
MSXOutFile was reading MAGIC by offsetting 4 bytes from the end; on 64 bit machines this
needs to be 8 bytes. Corrected this by using sizeof(INT4) instead of statically defined
RECORDSIZE.

As of this revision, MSX compiles and executes on Linux and produces results for test problem
that are identical to those produced on WIN32.