MSX changes:  - added code to MSXfuncs_run() and MSXfuncs_load() to convert the  argument to wide chars, cause that's what the windows  system calls expect.  Fixed bug that caused compilation option  to fail on VC.  - fixed bug in automatic C code generation to use pow()  instead of ^ (which was what was intended anyway).  - changed gcc compiler command string to use -O3 optimization  level, as this was proven safe on test cases and results in  about 10% performance improvement.
Changed the msx makefiles to also specify -O3 for gcc builds.

Updates to *.vcproj files for MSVS builds

Collected all of the text "Readme" files into the trunk/doc
directory, for safe keeping and eventual distribution through
the release process.  Previously these were only located in the
tagged release subdirs.

Updates to the test script and results that document output as
of this build.